Change values to double

+++ unit-testing-using-dotnet-test/AddService/AddService.cs
@@ -2,7 +2,7 @@
 
 namespace Add.Service;
 public class AddService {
-  public double Add(int num1, int num2) {
+  public double Add(double num1, double num2) {
     
     return num1 + num2;
